+++ pylangacq/_punctuation_marks.py
@@ -1,0 +1,2 @@
+# See https://www.talkbank.org/manuals/CHAT.pdf for the punctuation set.
+_PUNCTUATION_MARKS = frozenset(",.;?!")